Create loanCalculation
This is the loan calculation I have so far (relies on the parameters in the top from the person in charge of majors and jobs). There is some issue I have to figure out with the calculation of the loan but I will come back to that later. 
Hopefully, others can use this to start off their code (setting up consistent variables).