Add files via upload
deleted previous testScores.java file. This new file takes input from the user in the form of their average tilt score, compares it against an upper and lower range array to find the correlating majorGrouping, then prompts the user to determine if they want to see the specific majors in their majorGrouping.